requirements.txt
pyannote.audio
python-dotenv
openai-whisper
